Session and SessionManager
This commit contains an initial implementation of the Session
class and a partial implementation of the SessionManager class.

The SessionManager has an Open function which is intended to
handle incoming WebSocket requests. Its responsibilty it to then
manage the connection for its lifetime. It manages this connection
through the Session class.

The SessionManager accepts a callback function which it invokes
with each received message. It also allows the caller to send a
message to a client based on their session ID.